refactor(styles): move index css to styles folder
